requirements.txt
google-generativeai
streamlit
requests
beautifulsoup4